// src/app/model/user.ts
export class user {
    iduser: number;
    idrol: number;
    firstname: string;
    secondname: string;
    surname: string;
    secondsurname: string;
    email: string;
    phone: string;
    username: string;
    dateborn: string;
    pass: string;
    token: string;
    rolname: string;
    dateadd: string;
    datemod: string;
    iduseradd: number;
    idusermod: number;
}